+++ gambler/source/Fold/Shortcut.hs
@@ -17,7 +17,7 @@ module Fold.Shortcut
 
     {- * Conversion -} fold, effectfulFold, nonemptyFold, shortcutNonemptyFold,
 
-    {- * Utilities -} demotivate,
+    {- * Utilities -} demotivate, duplicate,
   )
   where