+++ gambler/source/Fold/Shortcut/Examples/Interesting.hs
@@ -10,14 +10,15 @@ module Fold.Shortcut.Examples.Interesting
 import Fold.Shortcut.Type
 
 import Control.Applicative (liftA2)
-import Data.Bool (Bool (False, True))
+import Data.Bool (Bool)
 import Data.Eq (Eq, (/=), (==))
 import Data.Functor (($>), (<&>))
 import Data.Maybe (Maybe (Just, Nothing))
-import Numeric.Natural (Natural)
-import Prelude ((-))
 import Fold.Shortcut.Conversion (fold)
 import Fold.Shortcut.Utilities (demotivate)
+import Numeric.Natural (Natural)
+import Prelude ((-))
+import Strict (isAlive, isDead)
 
 import qualified Fold.Pure.Examples.Interesting as Fold
 
@@ -26,8 +27,7 @@ null :: ShortcutFold a Bool
 null = ShortcutFold
   { initial = Alive Tenacious ()
   , step = \() _ -> Dead ()
-  , extractLive = \() -> True
-  , extractDead = \() -> False
+  , extract = isAlive
   }
 
 {-| 'True' if all inputs are 'True' (tenacious) -}
@@ -35,8 +35,7 @@ and :: ShortcutFold Bool Bool
 and = ShortcutFold
   { initial = Alive Tenacious ()
   , step = \_ a -> if a then Alive Tenacious () else Dead ()
-  , extractLive = \() -> True
-  , extractDead = \() -> False
+  , extract = isAlive
   }
 
 {-| 'True' if any input is 'True' (tenacious) -}
@@ -44,8 +43,7 @@ or :: ShortcutFold Bool Bool
 or = ShortcutFold
   { initial = Alive Tenacious ()
   , step = \() a -> if a then Dead () else Alive Tenacious ()
-  , extractLive = \() -> False
-  , extractDead = \() -> True
+  , extract = isDead
   }
 
 {-| 'True' if all inputs satisfy the predicate (tenacious) -}
@@ -53,8 +51,7 @@ all :: (a -> Bool) -> ShortcutFold a Bool
 all predicate = ShortcutFold
   { initial = Alive Tenacious ()
   , step = \() a -> if predicate a then Alive Tenacious () else Dead ()
-  , extractLive = \() -> True
-  , extractDead = \() -> False
+  , extract = isAlive
   }
 
 {-| 'True' if any input satisfies the predicate (tenacious) -}
@@ -62,8 +59,7 @@ any :: (a -> Bool) -> ShortcutFold a Bool
 any predicate = ShortcutFold
   { initial = Alive Tenacious ()
   , step = \_ a -> if predicate a then Dead () else Alive Tenacious ()
-  , extractLive = \() -> False
-  , extractDead = \() -> True
+  , extract = isDead
   }
 
 {-| 'True' if any input is equal to the given value (tenacious) -}
@@ -79,8 +75,7 @@ find :: (a -> Bool) -> ShortcutFold a (Maybe a)
 find ok = ShortcutFold
     { initial = Alive Tenacious ()
     , step = \() a -> if ok a then Dead a else Alive Tenacious ()
-    , extractDead = Just
-    , extractLive = \() -> Nothing
+    , extract = \v -> case v of { Dead x -> Just x; _ -> Nothing }
     }
 
 {-| The /n/th input, where n=0 is the first input, if the index is in
@@ -89,8 +84,7 @@ index :: Natural -> ShortcutFold a (Maybe a)
 index i = ShortcutFold
     { initial = Alive Tenacious i
     , step = \i' a -> if i' == 0 then Dead a else Alive Tenacious (i' - 1)
-    , extractDead = Just
-    , extractLive = \_ -> Nothing
+    , extract = \v -> case v of { Dead x -> Just x; _ -> Nothing }
     }
 
 {-| The index of the first input that matches the given value, if any
@@ -113,6 +107,5 @@ lookup :: Eq a => a -> ShortcutFold (a, b) (Maybe b)
 lookup a0 = ShortcutFold
     { initial = Alive Tenacious ()
     , step = \() (a, b) -> if a == a0 then Dead b else Alive Tenacious ()
-    , extractLive = \() -> Nothing
-    , extractDead = Just
+    , extract = \v -> case v of { Dead x -> Just x; _ -> Nothing }
     }