+++ gambler/source/Fold/Shortcut/Utilities.hs
@@ -8,12 +8,23 @@ import qualified Strict
 
 {-| Causes a shortcut fold to stop once it becomes ambivalent -}
 demotivate :: ShortcutFold a b -> ShortcutFold a b
-demotivate ShortcutFold{ initial, step, extractDead, extractLive } =
+demotivate ShortcutFold{ initial, step, extract } =
   ShortcutFold
     { initial = willSave initial
     , step = \x a -> willSave (step x a)
-    , extractDead = \e -> case e of
-          Strict.Left x -> extractDead x
-          Strict.Right x -> extractLive x
-    , extractLive = extractLive
+    , extract = \v -> case v of
+        Dead e -> case e of
+          Strict.Left x -> extract (Dead x)
+          Strict.Right x -> extract (Alive Ambivalent x)
+        Alive w x -> extract (Alive w x)
+    }
+
+{-| Allows to continue feeding a fold even after passing it to a function
+that closes it -}
+duplicate :: ShortcutFold a b -> ShortcutFold a (ShortcutFold a b)
+duplicate ShortcutFold{ initial, step, extract } =
+  ShortcutFold
+    { initial
+    , step
+    , extract = \v -> ShortcutFold{ initial = v, step, extract }
     }