+++ gambler/source/Fold/ShortcutNonempty/Examples/Interesting.hs
@@ -6,7 +6,6 @@ module Fold.ShortcutNonempty.Examples.Interesting
 
 import Fold.ShortcutNonempty.Type
 
-import Data.Function (id)
 import Data.Void (absurd)
 
 {-| The first input (tenacious) -}
@@ -14,6 +13,5 @@ first :: ShortcutNonemptyFold a a
 first = ShortcutNonemptyFold
   { initial = Dead
   , step = absurd
-  , extractDead = id
-  , extractLive = absurd
+  , extract = \v -> case v of { Dead x -> x }
   }