+++ gambler/source/Fold/ShortcutNonempty/Type.hs
@@ -19,43 +19,32 @@ import qualified Strict
 data ShortcutNonemptyFold a b = forall x y. ShortcutNonemptyFold
     { initial :: a -> Vitality x y
     , step :: y -> a -> Vitality x y
-    , extractDead :: x -> b
-    , extractLive :: y -> b
+    , extract :: Vitality x y -> b
     }
 
 instance Functor (ShortcutNonemptyFold a) where
-    fmap f ShortcutNonemptyFold{ step, initial, extractDead, extractLive } =
-        ShortcutNonemptyFold
-          { initial
-          , step
-          , extractDead = \x -> f (extractDead x)
-          , extractLive = \x -> f (extractLive x)
-          }
+    fmap f ShortcutNonemptyFold{ step, initial, extract } =
+        ShortcutNonemptyFold{ initial, step, extract = \x -> f (extract x) }
 
 instance Applicative (ShortcutNonemptyFold a) where
     pure b = ShortcutNonemptyFold
         { initial = \_ -> Dead ()
         , step = absurd
-        , extractDead = \() -> b
-        , extractLive = absurd
+        , extract = \e -> case e of { Dead () -> b }
         }
 
     (<*>)
-        ShortcutNonemptyFold{ initial = initialL, step = stepL, extractDead = extractDeadL, extractLive = extractLiveL }
-        ShortcutNonemptyFold{ initial = initialR, step = stepR, extractDead = extractDeadR, extractLive = extractLiveR } =
+        ShortcutNonemptyFold{ initial = initialL, step = stepL, extract = extractL }
+        ShortcutNonemptyFold{ initial = initialR, step = stepR, extract = extractR } =
           ShortcutNonemptyFold
             { initial = \a -> Strict.vitality2 (initialL a) (initialR a)
             , step = \(Strict.Tuple2 xL xR) a -> Strict.vitality2
                 (Strict.unlessDead (\x -> stepL x a) xL)
                 (Strict.unlessDead (\x -> stepR x a) xR)
-            , extractDead = extract
-            , extractLive = extract
+            , extract = \e -> case e of { Dead x -> ex x; Alive _ x -> ex x }
             }
           where
-            extract(Strict.Tuple2 xL xR) = f x
-              where
-                f = case xL of { Dead a -> extractDeadL a; Alive _ b -> extractLiveL b }
-                x = case xR of { Dead a -> extractDeadR a; Alive _ b -> extractLiveR b }
+            ex (Strict.Tuple2 xL xR) = (extractL xL) (extractR xR)
 
 instance Semigroup b => Semigroup (ShortcutNonemptyFold a b) where
     (<>) = liftA2 (<>)