+++ gambler/test/Spec/Nonempty.hs
@@ -4,9 +4,13 @@ import Fold.Nonempty
 
 import Test.Hspec
 
+import Data.Function ((&), flip)
 import Data.List.NonEmpty (NonEmpty ((:|)))
 import Positive (Positive)
 import Prelude (String, Integer)
+
+import qualified Data.List as List
+import qualified Fold.Pure as Pure
 
 spec :: SpecWith ()
 spec = describe "NonemptyFold" do
@@ -41,3 +45,12 @@ spec = describe "NonemptyFold" do
             run sum [1,2,5] `shouldBe` (8 :: Positive)
         it "product works with Positive" do
             run product [1,2,5] `shouldBe` (10 :: Positive)
+
+    describe "duplicate" do
+        it "lets a fold run in two phases" do
+            let a, c :: NonEmpty Integer
+                b :: [Integer]
+                a = [1..3]
+                b = [4..6]
+                c = [1..6]
+            (sum & duplicate & flip run a & flip Pure.run b) `shouldBe` (List.sum c)